Edit pytest to take optional environment config and add new test
Edit pytests so that custom environment variables can be set for each node
for each test. These environment variables are propagated to pytest cache
with a fixture and can be applied as an optional argument for cluster bring
up function.

Add new test that takes mixer delay as a env value and tests if the setting
works